Restore signed TOTP seconds offset semantics

+++ deepwell/src/config/object.rs
@@ -117,8 +117,8 @@ pub struct Config {
     /// Length in seconds that each TOTP lasts.
     pub totp_time_step: u64,
 
-    /// How many seconds of clock-skew leniency should be allowed for TOTP.
-    pub totp_time_skew: u32,
+    /// Signed seconds added to the server timestamp before TOTP verification.
+    pub totp_time_skew: i64,
 
     /// The number of job workers to run in this process.
     pub job_workers: NonZeroU16,
